Refactoring base tests into oklog-core-test project
