+++ core-test/build.gradle
@@ -1,0 +1,33 @@
+apply plugin: 'com.android.library'
+
+android {
+    compileSdkVersion 23
+    buildToolsVersion '23.0.2'
+
+    defaultConfig {
+        minSdkVersion 15
+        targetSdkVersion 23
+        versionCode Integer.parseInt(VERSION_CODE)
+        versionName VERSION_NAME
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+
+    lintOptions {
+        checkReleaseBuilds false
+        textReport true
+        lintConfig file('../config/check/lint/lint.xml')
+    }
+}
+
+dependencies {
+    compile project(':oklog-core')
+
+    compile "junit:junit:$junitVersion"
+    compile "org.mockito:mockito-core:$mockitoVersion"
+}
