+++ oklog3/src/test/java/com/github/simonpercic/oklog3/LogDataInterceptorUnitTest.java
@@ -1,19 +1,13 @@
 package com.github.simonpercic.oklog3;
 
-import com.github.simonpercic.oklog.core.LogDataBuilder;
-import com.github.simonpercic.oklog.core.LogDataBuilder.HeaderData;
 import com.github.simonpercic.oklog.core.LogManager;
-import com.github.simonpercic.oklog.core.StringUtils;
+import com.github.simonpercic.oklog.coretest.BaseLogDataInterceptorUnitTest;
 
-import org.junit.Before;
 import org.junit.Rule;
-import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
 import java.io.IOException;
-import java.util.List;
 import java.util.Locale;
 
 import okhttp3.HttpUrl;
@@ -25,12 +19,6 @@ import okhttp3.Response;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import okio.Buffer;
-import okio.ByteString;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
 
 /**
  * LogDataInterceptor unit test. Inspired by: https://github.com/square/okhttp/tree/master/okhttp-logging-interceptor.
@@ -38,24 +26,16 @@ import static org.junit.Assert.fail;
  * @author Simon Percic <a href="https://github.com/simonpercic">https://github.com/simonpercic</a>
  */
 @RunWith(MockitoJUnitRunner.class)
-public class LogDataInterceptorUnitTest {
+public class LogDataInterceptorUnitTest extends BaseLogDataInterceptorUnitTest<MockResponse, Request> {
 
-    private static final MediaType PLAIN = MediaType.parse("text/plain; charset=utf-8");
-    private static final String CONTENT_TYPE = "Content-Type";
-    private static final String GET = "GET";
-    private static final String MESSAGE_OK = "OK";
-    private static final int RESPONSE_CODE_OK = 200;
-
-    @Mock LogManager applicationLogManager;
-    @Mock LogManager networkLogManager;
+    private static final MediaType PLAIN = MediaType.parse(PLAIN_STRING);
 
     @Rule public final MockWebServer server = new MockWebServer();
 
     private OkHttpClient client;
-    private String host;
     private HttpUrl url;
 
-    @Before public void setUp() {
+    @Override protected String createClient(LogManager applicationLogManager, LogManager networkLogManager) {
         OkLogInterceptor applicationInterceptor = new OkLogInterceptor(applicationLogManager);
         OkLogInterceptor networkInterceptor = new OkLogInterceptor(networkLogManager);
 
@@ -64,302 +44,58 @@ public class LogDataInterceptorUnitTest {
                 .addNetworkInterceptor(networkInterceptor)
                 .build();
 
-        host = String.format(Locale.US, "%s:%d", server.getHostName(), server.getPort());
         url = server.url("/");
+
+        return String.format(Locale.US, "%s:%d", server.getHostName(), server.getPort());
     }
 
-    @Test public void bodyGet() throws IOException {
-        server.enqueue(new MockResponse());
-        Response response = client.newCall(request().build()).execute();
-        response.body().close();
-
-        LogDataBuilder expectedValue = expectedLogData()
-                .requestMethod(GET)
-                .requestBodyState(LogDataBuilder.NO_BODY)
-                .responseCode(RESPONSE_CODE_OK)
-                .responseMessage(MESSAGE_OK);
-
-        LogDataBuilder appValue = TestUtils.getLogData(applicationLogManager);
-
-        TestUtils.assertData(expectedValue, appValue);
-        assertNull(appValue.getRequestHeaders());
-
-        LogDataBuilder networkValue = TestUtils.getLogData(networkLogManager);
-        TestUtils.assertData(expectedValue, networkValue);
-        assertRequestHeaders(networkValue.getRequestHeaders());
+    @Override protected void newCall(MockResponse mockResponse) throws IOException {
+        newCall(request().build(), mockResponse);
     }
 
-    @Test public void bodyGet204() throws IOException {
-        bodyGetNoBody(204);
-    }
-
-    @Test public void bodyGet205() throws IOException {
-        bodyGetNoBody(205);
-    }
-
-    private void bodyGetNoBody(int code) throws IOException {
-        server.enqueue(new MockResponse()
-                .setStatus("HTTP/1.1 " + code + " No Content"));
-        Response response = client.newCall(request().build()).execute();
-        response.body().close();
-
-        LogDataBuilder expectedValue = expectedLogData()
-                .requestMethod(GET)
-                .requestBodyState(LogDataBuilder.NO_BODY)
-                .responseCode(code)
-                .responseMessage("No Content");
-
-        LogDataBuilder appValue = TestUtils.getLogData(applicationLogManager);
-        TestUtils.assertData(expectedValue, appValue);
-        assertNull(appValue.getRequestHeaders());
-
-        LogDataBuilder networkValue = TestUtils.getLogData(networkLogManager);
-        TestUtils.assertData(expectedValue, networkValue);
-        assertRequestHeaders(networkValue.getRequestHeaders());
-    }
-
-    @Test public void bodyPost() throws IOException {
-        server.enqueue(new MockResponse());
-        Request request = request().post(RequestBody.create(PLAIN, "Hi?")).build();
+    @Override protected void newCall(Request request, MockResponse mockResponse) throws IOException {
+        server.enqueue(mockResponse);
         Response response = client.newCall(request).execute();
         response.body().close();
-
-        LogDataBuilder expectedValue = expectedLogData()
-                .requestMethod("POST")
-                .requestContentType(PLAIN.toString())
-                .requestContentLength(3)
-                .requestBody("Hi?")
-                .responseCode(RESPONSE_CODE_OK)
-                .responseMessage(MESSAGE_OK);
-
-        LogDataBuilder appValue = TestUtils.getLogData(applicationLogManager);
-        TestUtils.assertData(expectedValue, appValue);
-        assertNull(appValue.getRequestHeaders());
-
-        LogDataBuilder networkValue = TestUtils.getLogData(networkLogManager);
-        TestUtils.assertData(expectedValue, appValue);
-        assertRequestHeaders(networkValue.getRequestHeaders());
     }
 
-    @Test public void bodyResponseBody() throws IOException {
-        server.enqueue(new MockResponse()
-                .setBody("Hello!")
-                .setHeader(CONTENT_TYPE, PLAIN));
-        Response response = client.newCall(request().build()).execute();
-        response.body().close();
-
-        LogDataBuilder expectedValue = expectedLogData()
-                .requestMethod(GET)
-                .requestBodyState(LogDataBuilder.NO_BODY)
-                .responseCode(RESPONSE_CODE_OK)
-                .responseMessage(MESSAGE_OK)
-                .responseBodySize(6)
-                .responseContentLength(6)
-                .responseBody("Hello!");
-
-        LogDataBuilder appValue = TestUtils.getLogData(applicationLogManager);
-        TestUtils.assertData(expectedValue, appValue);
-        assertNull(appValue.getRequestHeaders());
-        assertHeaderData(appValue.getResponseHeaders(), CONTENT_TYPE, new PredicateEquals<>(PLAIN.toString()));
-
-        LogDataBuilder networkValue = TestUtils.getLogData(networkLogManager);
-        TestUtils.assertData(expectedValue, networkValue);
-        assertRequestHeaders(networkValue.getRequestHeaders());
-        assertHeaderData(appValue.getResponseHeaders(), CONTENT_TYPE, new PredicateEquals<>(PLAIN.toString()));
+    @Override protected String getUrlString() {
+        return url.toString();
     }
 
-    @Test public void bodyResponseBodyChunked() throws IOException {
-        server.enqueue(new MockResponse()
-                .setChunkedBody("Hello!", 2)
-                .setHeader(CONTENT_TYPE, PLAIN));
-        Response response = client.newCall(request().build()).execute();
-        response.body().close();
-
-        LogDataBuilder expectedValue = expectedLogData()
-                .requestMethod(GET)
-                .requestBodyState(LogDataBuilder.NO_BODY)
-                .responseCode(RESPONSE_CODE_OK)
-                .responseMessage(MESSAGE_OK)
-                .responseBodySize(6)
-                .responseContentLength(-1)
-                .responseBody("Hello!");
-
-        LogDataBuilder appValue = TestUtils.getLogData(applicationLogManager);
-        TestUtils.assertData(expectedValue, appValue);
-        assertNull(appValue.getRequestHeaders());
-        assertHeaderData(appValue.getResponseHeaders(), "Transfer-encoding", new PredicateEquals<>("chunked"));
-        assertHeaderData(appValue.getResponseHeaders(), CONTENT_TYPE, new PredicateEquals<>(PLAIN.toString()));
-
-        LogDataBuilder networkValue = TestUtils.getLogData(networkLogManager);
-        TestUtils.assertData(expectedValue, networkValue);
-        assertRequestHeaders(networkValue.getRequestHeaders());
-        assertHeaderData(networkValue.getResponseHeaders(), "Transfer-encoding", new PredicateEquals<>("chunked"));
-        assertHeaderData(networkValue.getResponseHeaders(), CONTENT_TYPE, new PredicateEquals<>(PLAIN.toString()));
+    @Override protected MockResponse createMockResponse() {
+        return new MockResponse();
     }
 
-    @Test public void bodyResponseNotIdentityEncoded() throws IOException {
-        server.enqueue(new MockResponse()
-                .setHeader("Content-Encoding", "gzip")
-                .setHeader(CONTENT_TYPE, PLAIN)
-                .setBody(new Buffer().write(ByteString.decodeBase64(
-                        "H4sIAAAAAAAAAPNIzcnJ11HwQKIAdyO+9hMAAAA="))));
-        Response response = client.newCall(request().build()).execute();
-        response.body().close();
-
-        LogDataBuilder expectedValue = expectedLogData()
-                .requestMethod(GET)
-                .requestBodyState(LogDataBuilder.NO_BODY)
-                .responseCode(RESPONSE_CODE_OK)
-                .responseMessage(MESSAGE_OK)
-                .responseBodySize(19)
-                .responseContentLength(-1)
-                .responseBody("Hello, Hello, Hello");
-
-        LogDataBuilder appValue = TestUtils.getLogData(applicationLogManager);
-        TestUtils.assertData(expectedValue, appValue);
-        assertNull(appValue.getRequestHeaders());
-        assertHeaderData(appValue.getResponseHeaders(), CONTENT_TYPE, new PredicateEquals<>(PLAIN.toString()));
-
-        expectedValue
-                .responseBodySize(0)
-                .responseContentLength(29)
-                .responseBodyState(LogDataBuilder.ENCODED_BODY)
-                .responseBody(null);
-
-        LogDataBuilder networkValue = TestUtils.getLogData(networkLogManager);
-        TestUtils.assertData(expectedValue, networkValue);
-        assertRequestHeaders(networkValue.getRequestHeaders());
-        assertHeaderData(networkValue.getResponseHeaders(), CONTENT_TYPE, new PredicateEquals<>(PLAIN.toString()));
-        assertHeaderData(networkValue.getResponseHeaders(), "Content-Encoding", new PredicateEquals<>("gzip"));
+    @Override protected MockResponse setMockResponseStatus(MockResponse mockResponse, String status) {
+        return mockResponse.setStatus(status);
     }
 
-    @Test public void bodyGetMalformedCharset() throws IOException {
-        String contentType = "text/html; charset=0";
-        server.enqueue(new MockResponse()
-                .setHeader(CONTENT_TYPE, contentType)
-                .setBody("Ignore This"));
-        Response response = client.newCall(request().build()).execute();
-        response.body().close();
-
-        LogDataBuilder expectedValue = expectedLogData()
-                .requestMethod(GET)
-                .requestBodyState(LogDataBuilder.NO_BODY)
-                .responseCode(RESPONSE_CODE_OK)
-                .responseMessage(MESSAGE_OK)
-                .responseBodySize(0)
-                .responseContentLength(11)
-                .responseBodyState(LogDataBuilder.CHARSET_MALFORMED);
-
-        LogDataBuilder appValue = TestUtils.getLogData(applicationLogManager);
-        TestUtils.assertData(expectedValue, appValue);
-        assertNull(appValue.getRequestHeaders());
-        assertHeaderData(appValue.getResponseHeaders(), CONTENT_TYPE, new PredicateEquals<>(contentType));
-
-        LogDataBuilder networkValue = TestUtils.getLogData(networkLogManager);
-        TestUtils.assertData(expectedValue, networkValue);
-        assertRequestHeaders(networkValue.getRequestHeaders());
-        assertHeaderData(networkValue.getResponseHeaders(), CONTENT_TYPE, new PredicateEquals<>(contentType));
+    @Override protected MockResponse setMockResponseBody(MockResponse mockResponse, String body) {
+        return mockResponse.setBody(body);
     }
 
-    @Test public void responseBodyIsBinary() throws IOException {
-        Buffer buffer = new Buffer();
-        buffer.writeUtf8CodePoint(0x89);
-        buffer.writeUtf8CodePoint(0x50);
-        buffer.writeUtf8CodePoint(0x4e);
-        buffer.writeUtf8CodePoint(0x47);
-        buffer.writeUtf8CodePoint(0x0d);
-        buffer.writeUtf8CodePoint(0x0a);
-        buffer.writeUtf8CodePoint(0x1a);
-        buffer.writeUtf8CodePoint(0x0a);
+    @Override protected MockResponse setMockResponseBody(MockResponse mockResponse, Buffer body) {
+        return mockResponse.setBody(body);
+    }
 
-        String contentType = "image/png; charset=utf-8";
-        server.enqueue(new MockResponse()
-                .setBody(buffer)
-                .setHeader(CONTENT_TYPE, contentType));
-        Response response = client.newCall(request().build()).execute();
-        response.body().close();
+    @Override protected MockResponse setMockResponsePlainContentTypeHeader(MockResponse mockResponse) {
+        return mockResponse.setHeader(CONTENT_TYPE, PLAIN);
+    }
 
-        LogDataBuilder expectedValue = expectedLogData()
-                .requestMethod(GET)
-                .requestBodyState(LogDataBuilder.NO_BODY)
-                .responseCode(RESPONSE_CODE_OK)
-                .responseMessage(MESSAGE_OK)
-                .responseBodySize(9)
-                .responseContentLength(9)
-                .responseBodyState(LogDataBuilder.BINARY_BODY);
+    @Override protected MockResponse setMockResponseHeader(MockResponse mockResponse, String name, Object value) {
+        return mockResponse.setHeader(name, value);
+    }
 
-        LogDataBuilder appValue = TestUtils.getLogData(applicationLogManager);
-        TestUtils.assertData(expectedValue, appValue);
-        assertNull(appValue.getRequestHeaders());
-        assertHeaderData(appValue.getResponseHeaders(), CONTENT_TYPE, new PredicateEquals<>(contentType));
+    @Override protected MockResponse setMockResponseChunked(MockResponse mockResponse, String body, int maxChunkSize) {
+        return mockResponse.setChunkedBody(body, maxChunkSize);
+    }
 
-        LogDataBuilder networkValue = TestUtils.getLogData(networkLogManager);
-        TestUtils.assertData(expectedValue, networkValue);
-        assertRequestHeaders(networkValue.getRequestHeaders());
-        assertHeaderData(networkValue.getResponseHeaders(), CONTENT_TYPE, new PredicateEquals<>(contentType));
+    @Override protected Request createPlainPostRequest(String body) {
+        return request().post(RequestBody.create(PLAIN, body)).build();
     }
 
     private Request.Builder request() {
         return new Request.Builder().url(url);
-    }
-
-    private LogDataBuilder expectedLogData() {
-        return new LogDataBuilder()
-                .requestUrl(url.toString())
-                .responseUrl(url.toString());
-    }
-
-    private void assertRequestHeaders(List<HeaderData> requestHeaders) {
-        assertHeaderData(requestHeaders, "Host", new PredicateEquals<>(host));
-        assertHeaderData(requestHeaders, "User-Agent", new PredicateStartsWith("okhttp"));
-    }
-
-    private void assertHeaderData(List<HeaderData> headers, String name, Predicate<String> assertPredicate) {
-        if (headers == null || headers.size() == 0) {
-            fail("Headers are empty");
-        }
-
-        if (StringUtils.isEmpty(name)) {
-            fail("Header name is empty");
-        }
-
-        for (HeaderData header : headers) {
-            if (name.equalsIgnoreCase(header.getName())) {
-                assertPredicate.assertValue(header.getValue());
-                return;
-            }
-        }
-
-        fail(String.format("No header with name: %s", name));
-    }
-
-    class PredicateStartsWith implements Predicate<String> {
-
-        final String expected;
-
-        PredicateStartsWith(String expected) {
-            this.expected = expected.toLowerCase(Locale.US);
-        }
-
-        @Override public void assertValue(String value) {
-            assertTrue(value.toLowerCase(Locale.US).startsWith(expected));
-        }
-    }
-
-    class PredicateEquals<T> implements Predicate<T> {
-
-        final T expected;
-
-        PredicateEquals(T expected) {
-            this.expected = expected;
-        }
-
-        @Override public void assertValue(T value) {
-            assertEquals(expected, value);
-        }
-    }
-
-    interface Predicate<T> {
-        void assertValue(T value);
     }
 }
